Update soda bread from latest test batch.
Notes:
Loaf still seems a bit dry, but overall, more "bready", which is good. Probably because using spread,
not actual butter.
May up the spread again on the next try.
Grinding the carroway made less of a random carroway experience.
Used the bread hook to mix and knead the loaf, seemed to work well.
Letting the loaf rest in the mixing bowl also pretty good idea. Outside of dough was sticky, but not terrible.
Second knead was mostly to firm up the loaf.